evoteapp/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1;python_version<"3.9"
Django==4.2.5
django-cors-headers==4.3.1
django-phonenumber-field==7.1.0
django-sendsms==0.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
httplib2==0.22.0
phonenumbers==8.13.22
Pillow==10.1.0
pip==23.2.1
PyJWT==2.8.0
pyotp==2.9.0
pyparsing==3.1.1
pytz==2023.3.post1
setuptools==56.0.0
six==1.16.0
sqlparse==0.4.4
textmagic==2.0.3
typing_extensions==4.8.0
twilio==8.11.0
python-dotenv==1.0.0